\name{phenorice-package}

\alias{phenorice-package}

\docType{package}

\title{
\packageTitle{phenorice}
}

\description{
phenorice model to detect rice from satellite reflectance data
}

\references{
Boschetti, M., L. Busettoa, G. Manfrona, A. Laborte, S. Asilo, S. Pazhanivelan & A Nelson, 2017. PhenoRice: A method for automatic extraction of spatio-temporal information on rice crops using satellite data time series. Remote Sensing of Environment 194: 347-365
}

